Refactored Views - AddToQueue view holds button for Media Picker - Player view holds artwork, song title, and album artist - as well as the PlayerControls view - PlayerControls view holds the play/pause, backward and forward buttons
